feat: refactor imports to use shapeshift models for preview and reconciliation services
